move variable to constants
