+++ src/util.ts
@@ -3,12 +3,6 @@ import path from 'path';
 
 import * as tc from '@actions/tool-cache';
 
-export const IS_WINDOWS = process.platform === 'win32';
-export const IS_LINUX = process.platform === 'linux';
-export const IS_MACOS = process.platform === 'darwin';
-export const PLATFORM = IS_WINDOWS ? 'windows' : process.platform;
-export const macOSJavaContentDir = 'Contents/Home';
-
 export function getTempDir() {
   let tempDirectory = process.env['RUNNER_TEMP'] || os.tmpdir();
 
@@ -43,5 +37,5 @@ export async function extractJdkFile(toolPath: string, extension?: string) {
 }
 
 export function getDownloadArchiveExtension() {
-  return IS_WINDOWS ? 'zip' : 'tar.gz';
+  return process.platform === 'win32' ? 'zip' : 'tar.gz';
 }
